requirements.txt
python-dotenv==1.1.0
openai==2.6.0
requests==2.31.0
pypdf==6.2.0
gradio==4.21.0
pydantic==2.12.0
python-pushover==0.5
psycopg2-binary==2.9.9

# Testing dependencies
pytest==8.0.0
pytest-mock==3.12.0
responses==0.25.0